Fix URL in ceremony script and make resumable
`prepare_ceremony` now accepts a number of command-line flags (see
`--help`) and also avoids duplicate work. The POT file will only be
downloaded if missing, circuits will only be compile if missing and
zkeys will only be generated if missing, such that the process can be
resumed. zkeys are by default now placed in a folder with the same name,
however this can be tweaked back to the old default by setting `--output
./build`